chg: dev: [ajaxReq,http,serializwer,endpoint] breakdown ``ajaxReq``
